+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+- Static HEEx children for client islands, passed to framework runtimes through the default slot/children channel.
+
 ## 0.1.2 - 2026-06-26
 
 ### Added